Catalina fixes -embedded a frame size value into CGFloat() which is a must on Catalina, but Xcode 13 can't detect this issue and builds the app. Build failed on Catalina with Xcode 12.4 -@StateObject isn't available on Catalina which Xcode 13 can't detect too. commented that line as we dont use related struct in production for the moment -Removed unnecessary print()

+++ HackinDROM/Views/Main View/HDUpdateView.swift
@@ -30,7 +30,7 @@ struct HDUpdateView: View {
             Color.red
                 .opacity(0.0)
             
-        }.frame(width: updatingPosition * 450)
+        }.frame(width: CGFloat(updatingPosition) * 450)
         .background(updatingColor.opacity(0.6))
         .contextMenu(menuItems: {
             Button("Cancel") {
